pyproject.toml
[project]
name = "podcaster"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.3",
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.5",
]
